bitreader: implement CachedBitReader and use it for huffman decoding
Implements a bitreader that caches 8 bytes in a 64 bit register and
provides an Iterator which reads the left most bit and shifts the
entire cache left by 1. This makes for a branchless faster bitreader.

To fill the cache 8 bytes are copied into the cache and shifted left by
the bit offset of the source bitreader. Because of this, only 57 bits
can be guaranteed to be valid.

This bitreader has to be manually refilled, which is ok for cases where
the maximum number of bits who are going to be read is predictable.

Because of the cache using a 64 bit register, this is enabled only when
compiling for 64 bit targets, since it would probably be slower on 32 bit.

This makes the overall decoding process about 3% faster.