bitreader: change strategy for CachedBitReader
Turns out that the amount of bits that the huffman decoder is going to read
is not as predictable. This threw things off.

What we do now instead is try to huffman decode, after which we check how many
bits were read and if the total amount is over 57 we refill the cache, giving us
57 fresh bits. We then retry the huffman decoding step.